requirements.txt
PyQt5==5.14.1
PyOpenGL==3.15